Fixed abundance p-value, Removed GSL dependence
DADA now uses R’s ppois function to calculate the poisson CDF, thereby
eliminating the previous dependence on the GSL C libraries. Of note, in
the process of testing this change, I found that the previous GSL
calculation was flawed (does not limit correctly when mu -> 0)!! This
must be fixed for the C version.