Added facility to set/get DADA options
There are a number of options that can be passed to the DADA algorithm.
set_dada_opt and get_dada_opt allow those options to be viewed/changed.

There is currently no consistency checking when setting options, this
would probably be good to add when closer to a finished product.