Graceful failure if OmegaS too stringent
If OmegaS is too stringent, it might be impossible to validly assign
significant singletons because (1) doubles aren’t precise enough or (2)
the summation has to go to far out to be computationally feasible.

These cases are now checked (in a semi-sane way), and if failed dada is
stopped with an appropriate error message.