Opts stored in dada return object
The score matrix and gap penalty used by dada is also now defined in
dada_opts, rather than passed in as an argument.

Eventually it might be useful to allow either method, but for now a
unified approach seems better. The error matrix and the self_consist
bool are now the only arguments to dada() other than the (list of)
uniques vector(s).