Pvals at cluster birth stored and reported
The pval when a cluster is formed is stored in the cluster and reported
in the object returned by dada() in the new return_val$clustering data
frame.

This pval does not necessarily indicate how confident we are in the
final cluster, it applies only at the birth of the cluster, but it does
give insight into where once might cutoff formation of new clusters
(i.e. where to set OmegaA).