Import for FastQ with multiple samples
The new function dereplicateMultiSampleFastqReads() reads in a fastq
file, and outputs uniques vectors for each sample. Samples are
identified by the subseq of the @id slot, which is the Fasta header for
each sequence.

The user must pass in the file name and the c(start,end) of the
sampleid subsequence.

Only works if sampleIDs are constant length (i.e. not
1,2…13,14,…100,101….)