Added filtering functionality
Added filterFastq(fn, fout,…) that implements much of the usearch
“fastq_filter” functionality. Checked it against the fastq_filter and
got an exact match.

The maxee filtering is not in, just the truncqual, trunclen, and
stripleft options. There is also a maxN filter as well (max number of
Ns).

Also included some utility functions to explore the error profile from
dada clusterings.