Quality score input/output
Quality scores can now be imported into the main dada algorithm, and an
output ($subqual) gives the relationship between the quality score of
uniques and the error rate.

Quality scores are not yet used in the clustering algorithm itself (eg.
by altering significance calculations).

uniques.cpp is longer used by the R package as of this update.