Error Models
DADA now does all estimation of error rates R side. Additionally, the
user can provide the estimation function if they so wish. The function
“threepiece” is provided as a recommended default for Illumina data
with quality scores.

Error rates are now specified in 16xN matrices, with rows corresponding
to the transition (1:A->A, 2:A->C, …, 5:C->A, …) and cols correspnoding
to the average quality score from QMIN to QMAX.

When USE_QUALS=FALSE, the err matris should be 16x1.