Cleaned up pair merging
Documented and cleaned up the mergePairs() function that is used to
merge forward and reverse reads together post-DADA-denoising.

Also changed the $subpos output of DADA to be a matrix of the
birth_subs of each cluster (row = t_ij, col=position). This will be
useful to identifiy problematic positions. Should make a utility plot
of the 1-away birth_subs.