Error function
The optional feature in which DADA calculates error rates from the
matrix of substitutions it infers now runs on a passed in function
parameter, err_function. err_function takes dada$trans_out, and returns
an error rate matrix. This is then used as the $err for the next loop
when self_consist = TRUE, or just reported as $err_out if not
self_consist looping.

Additionaly, documentation and naming consistency was improved.

isBimera is now provided as part of the package.

All dependence on seqinr was removed.

Time to merge back into main branch.