Integer score matrix/gap penalty
The score matrix and gap penalty are now coerced to integers and used
that way throughout the C side. This is a modest performance gain.

Also, the default BAND_SIZE was reduced to 16, as this had essentially
no impact on the test data. Should look at this (and KMER_SIZE) on 454
data independently though.