Final consensus
There is now a FINAL_CONSENSUS option in DADA to return a consensus
sequence for each cluster rather than the cluster center (the sequence
which started the cluster).

This is FALSE by default, and its not clear that doing a consensus is
an improvement, and might be the opposite. But its an option now in any
case.