Major memory gains: clean up
This commit cleans up the major memory gains from freeing Sub objects
associated with raws that did not join, or later left, a cluster.

There is a small peformance penalty, because if those raws later join
(or rejoin) that cluster, the Sub object is recomputed, but the memory
savings can be huge so the tradeoff is worthwhile.