VECTORIZED_ALIGNMENT
The new vectorized aligner is now on by default
(VECTORIZED_ALIGNMENT=TRUE). This should not have any affect on the
output of dada(…). Turning the flag off reverts to the previous aligner.

The vectorized aligner uses SIMD instructions to accelerate the
computation of the Needleman-Wunsch dynamic programming matrix.

However there is an overhead cost. The vectorized aligner is about 2x
faster for the default BAND_SIZE=16, but it becomes slower for
BAND_SIZE<=8. For unbanded alignments it is about 7x faster.