Removed Fams and Singletons
This is a major change. The fam structure, in which sequences that
differed only by indels to one another were joined together for the
purposes of calculating the abundance pvalue, was completely stripped
out. Bi’s are not simply containers for Raws.

The decaying singleton p-value code was also removed.

Semi-miraculously, it ran mostly correclty right away.

Need to clean up a bit more, and then test the differences introduced
here.