Further QMAX fixes and NULL initialization
I propagated through some more QMAX changes (mostly cosmetic).

There is also an added functionality, an err=NULL can be passed to the
dada(…) function if selfConsist=TRUE, in which case the maximum possile
error rates are estimated from the data as an initialization.

This corresponds to placing all sequences in one partition, as
described in the DADA1 paper (Rosen et al 2012).