getUniques update
Added flags to getUniques to collapse (or not) duplicates, and report
(or not) duplicates.

Reviewed its usage, and cleaned up a couple edge cases that happen when
functions using getUniques are called on input with duplicates
(shouldn’t happen usually, but can occasionally crop up due to unusual
actions like the user trimming sequences manually).