New chimera method for sequence tables
New method does sample-wise chimera flagging, and then “votes”, with
sequences flagged in >X% of the samples in which they are present
removed from the dataset as chimeras.

C_table_chimera implements this for now. Need to continue testing and
roll into R interface. Right now looks like an improvement though,
especially in getting difficult to call abundant sequences that
could-be-real/could-be-chimeras correct.