Modify verbose behavior in dada(...)
verbose=FALSE is now silence, and verbose=TRUE is normal output.
Default was changed to TRUE, so the default behavior of the function
remains the same. The debug level output is now achieved by verbose=2,
instead of verbose=TRUE as before. This moves the behavior more in line
with other functions in the package. Fixes #568