Added DETECT_SINGLETONS dadaOpt
FALSE by default, but if set to TRUE it removes the “conditional” in
the calculation of probabilties used in the core dada algorithm, which
effectively discounts the first read of any novel sequence. In
practice, this means that singletons will now be detected if they are
adjudged too unlikely to be produced under the error model, and also
increases sensitivity of detecting other low abundance sequences, i.e.
those present in just 2/3/4 reads.

In testing, this increases sensitivity to low frequency variants, but
lets a lot more junk into the final output. It is not recommended for
general purpose usage (hence the default of FALSE), but can be useful
for those who understand the high false-positive rate of singletons in
real data and want them anyway. It will also be useful as a potentially
recommended option for ultra-high-quality reads (e.g. Loop Genomics
consensus reads, and perhaps PacBio CCS in at least some circumstances)
where error rates are quire low and very well behaved.