Keep stuff after the species names
The stuff after the common species name is either a number, or a strain, or whatever taxonomists put in there. I don't know how many of them would be in my data if we weren't removing them, because they all got removed :).

The change is for consideration only, do close this if you think it's a step in a worse direction.
```
> ids=c("AJ294429.1.1559 Desulfofundulus thermobenzoicus", "CP012881.1458553.1460106 Vibrio vulnificus NBRC 15645 = ATCC 27562")
> gsub("^\\S*\\s\\S*\\s", "", ids)
[1] "thermobenzoicus"                    "vulnificus NBRC 15645 = ATCC 27562"
```

Found when investigating a species of genus "Candidatus Nitrocosmicus" that got assigned to a species "Nitrocosmicus" in my data. This change won't necessarily fix it, and I have since seen that you're removing the word Candidatus from the reference as a workaround.