derepFastq now handles zero-length reads in the input
Those reads are excluded from the $uniques and $quals parts of the
`derep-class` object, and have `NA` entries in the num-reads-long $map
vector. A warning is thrown when zero-length reads are encountered
though.

Downstream integration with other functions including `combineDereps2`
and `mergePairs` seems to work as expected.

At the moment this behavior is flagged by the default TRUE
`handle.zerolen` parameter in `derepFastq` and `qtables2`. This is just
for testing convenience, and should probably be removed before the next
release.

Fixes #1046
Fixes #1027